refactor: remove common package
