+++ package.json
@@ -9,17 +9,13 @@
   ],
   "sideEffects": false,
   "scripts": {
-    "common": "yarn workspace @whe/common",
-    "start:common": "yarn common start",
-    "build:common": "yarn common build",
-    "build:dep:common": "yarn build:common && mkdir -p dist && mv packages/common/dist dist/common",
-    "build:storybook": "yarn common build:storybook",
-    "test:common": "yarn common test",
-    "release:common": "yarn --cwd packages/common run standard-version",
     "blog": "yarn workspace @whe/blog",
     "start:blog": "yarn blog dev",
+    "storybook": "yarn blog start:storybook",
+    "build:storybook": "yarn blog build:storybook",
     "build:blog": "yarn blog build",
     "build:dep:blog": "yarn build:blog && mkdir -p dist/blog && cp -r packages/blog/{.next,public,package.json,next.config.js,next-i18next.config.js} dist/blog",
+    "build:dep:storybook": "yarn build:storybook && mkdir -p dist && mv packages/blog/build dist/storybook",
     "test:blog": "yarn blog test",
     "release:blog": "yarn --cwd packages/blog run standard-version",
     "api": "yarn workspace @whe/api",
@@ -41,11 +37,11 @@
     "postInstall": "patch-package"
   },
   "devDependencies": {
-    "@commitlint/cli": "16.2.3",
-    "@commitlint/config-conventional": "16.2.1",
+    "@commitlint/cli": "16.2.4",
+    "@commitlint/config-conventional": "16.2.4",
     "husky": "7.0.4",
-    "lint-staged": "12.4.0",
-    "npm-check-updates": "12.5.9",
+    "lint-staged": "12.4.1",
+    "npm-check-updates": "12.5.11",
     "npm-run-all": "4.1.5",
     "patch-package": "6.4.7",
     "prettier": "2.6.2",
