+++ packages/api/package.json
@@ -12,13 +12,13 @@
   },
   "devDependencies": {
     "@babel/eslint-parser": "7.17.0",
-    "@types/jest": "27.4.1",
-    "jest": "28.0.1"
+    "@types/jest": "27.5.0",
+    "jest": "28.1.0"
   },
   "dependencies": {
-    "@strapi/plugin-i18n": "4.1.8",
-    "@strapi/plugin-users-permissions": "4.1.8",
-    "@strapi/strapi": "4.1.8",
+    "@strapi/plugin-i18n": "4.1.9",
+    "@strapi/plugin-users-permissions": "4.1.9",
+    "@strapi/strapi": "4.1.9",
     "pg": "8.7.3",
     "strapi-plugin-ckeditor5": "2.1.1-rc.1"
   },
