+++ packages/blog/package.json
@@ -3,48 +3,63 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "dev": "NODE_OPTIONS='--inspect' next dev",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "jest --passWithNoTests"
+    "test": "jest --passWithNoTests",
+    "start:storybook": "start-storybook -p 9000",
+    "clean:storybook": "rm -rf ./build",
+    "build:storybook": "yarn clean:storybook && yarn build-storybook -o build",
+    "gen:theme-typings": "chakra-cli tokens ./src/theme/chakra.theme.ts",
+    "postinstall": "yarn gen:theme-typings"
   },
   "sideEffects": false,
   "dependencies": {
+    "@chakra-ui/icons": "1.1.7",
     "@chakra-ui/react": "2.0.0-next.0",
     "@chakra-ui/theme-tools": "1.3.6",
     "@emotion/react": "11.9.0",
     "@emotion/styled": "11.8.1",
-    "@fontsource/lora": "4.5.6",
+    "@fontsource/lora": "4.5.9",
     "@fontsource/roboto": "4.5.5",
     "@fontsource/spartan": "4.5.8",
-    "@whe/common": "*",
-    "axios": "0.27.1",
+    "axios": "0.27.2",
     "date-fns": "2.28.0",
-    "framer-motion": "6.3.2",
-    "i18next": "21.6.16",
+    "framer-motion": "6.3.3",
+    "i18next": "21.7.1",
     "ky": "0.30.0",
-    "next": "12.1.5",
+    "next": "12.1.6",
     "next-i18next": "11.0.0",
     "next-secure-headers": "2.2.0",
     "nextjs-progressbar": "0.0.14",
-    "react": "18.0.0",
-    "react-dom": "18.0.0",
-    "react-i18next": "11.16.7",
-    "react-query": "3.38.0",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-i18next": "11.16.9",
+    "react-query": "3.39.0",
     "schema-dts": "1.1.0",
     "slugify": "1.6.5"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "12.1.5",
+    "@chakra-ui/cli": "1.9.0",
+    "@chakra-ui/storybook-addon": "1.0.3",
+    "@next/bundle-analyzer": "12.1.6",
+    "@storybook/addon-actions": "6.4.22",
+    "@storybook/addon-essentials": "6.4.22",
+    "@storybook/addon-links": "6.4.22",
+    "@storybook/builder-webpack5": "6.5.0-alpha.22",
+    "@storybook/manager-webpack5": "6.5.0-alpha.22",
+    "@storybook/node-logger": "6.4.22",
+    "@storybook/preset-create-react-app": "4.1.0",
+    "@storybook/react": "6.4.22",
     "@svgr/webpack": "6.2.1",
-    "@types/jest": "27.4.1",
-    "@types/node": "17.0.27",
-    "@types/react": "18.0.7",
-    "@typescript-eslint/eslint-plugin": "5.21.0",
-    "@typescript-eslint/parser": "5.21.0",
-    "eslint": "8.14.0",
-    "eslint-config-next": "12.1.5",
+    "@types/jest": "27.5.0",
+    "@types/node": "17.0.31",
+    "@types/react": "18.0.9",
+    "@typescript-eslint/eslint-plugin": "5.22.0",
+    "@typescript-eslint/parser": "5.22.0",
+    "eslint": "8.15.0",
+    "eslint-config-next": "12.1.6",
     "eslint-config-prettier": "8.5.0",
     "eslint-import-resolver-node": "0.3.6",
     "eslint-import-resolver-typescript": "2.7.1",
@@ -54,13 +69,16 @@
     "eslint-plugin-jsx-a11y": "6.5.1",
     "eslint-plugin-prettier": "4.0.0",
     "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.4.0",
+    "eslint-plugin-react-hooks": "4.5.0",
     "eslint-plugin-regexp": "1.7.0",
-    "eslint-plugin-testing-library": "5.3.1",
-    "jest": "28.0.1",
+    "eslint-plugin-storybook": "0.5.11",
+    "eslint-plugin-testing-library": "5.4.0",
+    "jest": "28.1.0",
     "patch-package": "6.4.7",
     "picocolors": "1.0.0",
-    "ts-jest": "27.1.4",
-    "typescript": "4.6.3"
+    "ts-jest": "28.0.2",
+    "ts-node": "10.7.0",
+    "typescript": "4.6.4",
+    "webpack": "5.72.0"
   }
 }
