+++ packages/blog/src/app-providers.tsx
@@ -1,8 +1,8 @@
 import { ChakraProvider } from '@chakra-ui/provider'
-import { emotionTheme } from '@whe/common'
 import { AppProps } from 'next/app'
 import { ReactNode, useState } from 'react'
 import { DefaultOptions, Hydrate, MutationCache, QueryCache, QueryClient, QueryClientProvider } from 'react-query'
+import emotionTheme from './themes/emotion.theme'
 
 type AppProvidersProps = {
   children: ReactNode
