+++ packages/blog/src/components/common/Footer/Footer.tsx
