+++ packages/blog/src/components/common/index.ts
@@ -1,5 +1,9 @@
+export * from './Footer/Footer'
 export * from './GlobalStyles'
+export * from './Header/Header'
 export * from './Metadata'
 export * from './NavLink'
+export * from './PostCard/PostCard'
+export * from './Tag/Tag'
 export * from './WheFooter'
 export * from './WheHeader'
