+++ packages/blog/src/components/home/HeaderImage/HeaderImage.tsx
@@ -1,9 +1,9 @@
 import { Box, Text, Flex, Divider } from '@chakra-ui/layout'
 import { useBreakpointValue } from '@chakra-ui/media-query'
-import { Tag } from '@whe/common'
 import format from 'date-fns/format'
 import { useTranslation } from 'next-i18next'
 import Image from 'next/image'
+import { Tag } from '../../common/Tag/Tag'
 
 interface HeaderImageProps {
   imgSrc: string
