+++ packages/blog/src/components/home/LastPosts/LastPosts.tsx
@@ -1,6 +1,6 @@
 import { Flex, Heading, SimpleGrid } from '@chakra-ui/layout'
-import { PostCard } from '@whe/common'
 import { useTranslation } from 'next-i18next'
+import { PostCard } from '../../common/PostCard/PostCard'
 
 interface LastPostsProps {
   posts: Post[]
