+++ packages/blog/src/pages/_document.tsx
@@ -1,6 +1,6 @@
 import { ColorModeScript } from '@chakra-ui/color-mode'
-import { emotionTheme } from '@whe/common'
 import NextDocument, { Html, Head, Main, NextScript } from 'next/document'
+import emotionTheme from '../themes/emotion.theme'
 
 export default class Document extends NextDocument {
   render() {
