+++ packages/blog/src/types.d/styled.d.ts
@@ -1,5 +1,5 @@
 import '@emotion/react'
-import { Theme as EmotionTheme } from '@whe/common'
+import { Theme as EmotionTheme } from '../themes/emotion.theme'
 
 declare module '@emotion/react' {
   // eslint-disable-next-line @typescript-eslint/no-empty-interface
