--- packages/common/.lintstagedrc
@@ -1,5 +1,0 @@
-{
-  "*.+(js|jsx|ts|tsx)": [
-    "eslint --fix"
-  ]
-}
