--- packages/common/.yarnrc
@@ -1,1 +1,0 @@
---run.modules-folder ../../node_modules
