--- packages/common/package.json
@@ -1,64 +1,0 @@
-{
-  "name": "@whe/common",
-  "version": "0.0.1",
-  "description": "Common shared components library",
-  "main": "src/index",
-  "types": "src/index",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "start": "start-storybook -p 9000",
-    "build": "yarn clean && yarn build-storybook -o build",
-    "test": "jest",
-    "lint": "eslint . --ext .ts,.tsx,.js,.jsx",
-    "clean": "rm -rf ./build",
-    "gen:theme-typings": "chakra-cli tokens ./src/theme/chakra.theme.ts",
-    "postinstall": "yarn gen:theme-typings"
-  },
-  "dependencies": {
-    "@chakra-ui/icons": "1.1.7",
-    "@chakra-ui/react": "2.0.0-next.0",
-    "@emotion/react": "11.9.0",
-    "@emotion/styled": "11.8.1",
-    "@fontsource/lora": "4.5.6",
-    "@fontsource/roboto": "4.5.5",
-    "@fontsource/spartan": "4.5.8",
-    "framer-motion": "6.3.2"
-  },
-  "devDependencies": {
-    "@chakra-ui/cli": "1.9.0",
-    "@chakra-ui/storybook-addon": "1.0.3",
-    "@storybook/addon-actions": "6.4.22",
-    "@storybook/addon-essentials": "6.4.22",
-    "@storybook/addon-links": "6.4.22",
-    "@storybook/builder-webpack5": "6.5.0-alpha.22",
-    "@storybook/manager-webpack5": "6.5.0-alpha.22",
-    "@storybook/node-logger": "6.4.22",
-    "@storybook/preset-create-react-app": "4.1.0",
-    "@storybook/react": "6.4.22",
-    "@testing-library/jest-dom": "5.16.4",
-    "@testing-library/react": "13.1.1",
-    "@testing-library/user-event": "14.1.1",
-    "@types/jest": "27.4.1",
-    "@typescript-eslint/eslint-plugin": "5.21.0",
-    "@typescript-eslint/parser": "5.21.0",
-    "eslint": "8.14.0",
-    "eslint-config-prettier": "8.5.0",
-    "eslint-import-resolver-typescript": "2.7.1",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jest": "26.1.5",
-    "eslint-plugin-jest-formatting": "3.1.0",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-prettier": "4.0.0",
-    "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.4.0",
-    "eslint-plugin-regexp": "1.7.0",
-    "eslint-plugin-storybook": "0.5.11",
-    "eslint-plugin-testing-library": "5.3.1",
-    "jest": "28.0.1",
-    "ts-jest": "27.1.4",
-    "ts-node": "10.7.0",
-    "typescript": "4.6.3",
-    "webpack": "5.72.0"
-  }
-}
