--- packages/common/src/components/index.ts
@@ -1,4 +1,0 @@
-export { Header } from './Header/Header'
-export { Footer } from './Footer/Footer'
-export { PostCard } from './PostCard/PostCard'
-export { Tag } from './Tag/Tag'
