--- packages/common/src/index.ts
@@ -1,7 +1,0 @@
-export * from './components'
-import chakraTheme from './themes/chakra.theme'
-import emotionTheme, { Theme as EmotionTheme } from './themes/emotion.theme'
-
-export { chakraTheme }
-export { emotionTheme }
-export type Theme = EmotionTheme
