--- packages/common/tsconfig.build.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "include": ["src"]
-}
