--- packages/common/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": [
-    "src",
-    "**/*.ts",
-    "**/*.tsx",
-    "**/*.js",
-    "**/*.jsx",
-    "**/*.mjs",
-    "**/*.json",
-    "../blog/src/components/Metadata"
-  ]
-}
